requirements.txt
jupyterlab
numpy >=1.23.5
matplotlib >=3.5.1
scipy >=1.9.3
scikit-image[data] >=0.20
Pillow >=2.1.0
torch >=2.0.1
torchvision
torchsummary
einops >=0.3.0
timm
scikit-learn>=0.18
pandas>=1.3.5
tqdm>=4.62.3
wandb
altair
cairosvg>=2.7.1
cairocffi>=1.4.2
vl-convert-python
opencv-python